dev: rename test modules/methods for unittest runner instead of nose
- nose 1.3.7 is very old and not using any advanced test features that
  would require a 3rd party test library like nose2 or pytest
- default pattern is test_* for runner to pick up TestCases
