+++ tests/xform_test_case/test_attribute_columns.py
@@ -3,7 +3,6 @@ Some tests for the new (v0.9) spec is properly implemented.
 """
 import codecs
 import os
-import unittest
 
 import pyxform
 
@@ -14,7 +13,7 @@ from tests.xform_test_case.base import XFormTestCase
 class AttributeColumnsTest(XFormTestCase):
     maxDiff = None
 
-    def runTest(self):
+    def test_conversion(self):
         filename = "attribute_columns_test.xlsx"
         self.get_file_path(filename)
         expected_output_path = os.path.join(
@@ -35,7 +34,3 @@ class AttributeColumnsTest(XFormTestCase):
         with codecs.open(expected_output_path, "rb", encoding="utf-8") as expected_file:
             with codecs.open(self.output_path, "rb", encoding="utf-8") as actual_file:
                 self.assertXFormEqual(expected_file.read(), actual_file.read())
-
-
-if __name__ == "__main__":
-    unittest.main()
